Asset loader test for Runner
New proto-prototype runner-game-graphics.html. It loads the assets.
Press spacebar to get a new asset on screen, space again to see it
move (a single step), space a third time to get rid of it, then on
to the next. It has Orbit controls so use mouse/trackpad to look
around scene.

GLTFLoader and three library files had to be moved to build/ folder
(some weird legacy structure issue) and script type switched to 'module'
for any of this to work.