Implemented Runner's generateCoins
Current parameters mean continuous generation of coins - no visible gaps between coin lines! This is because a proper game loop has yet to be added, so there is no deltaTime.

Coins may generate above or below the fallen tree branch obstacle that is part of the trail not the obstacles. Considering splitting this off into an obstacle so it behaves consistently.